[comm] an initial prototype for hybrid server-driven communicator
