+++ examples/resources/config_hybrid/mnist/server.yaml
@@ -1,0 +1,78 @@
+client_configs:
+  train_configs:
+    # Local trainer
+    trainer: "VanillaTrainer"
+    mode: "step"
+    num_local_steps: 100
+    optim: "Adam"
+    optim_args:
+      lr: 0.001
+    # Loss function
+    loss_fn_path: "./resources/loss/celoss.py"
+    loss_fn_name: "CELoss"
+    # Client validation
+    do_validation: True
+    do_pre_validation: True
+    metric_path: "./resources/metric/acc.py"
+    metric_name: "accuracy"
+    # Differential privacy
+    use_dp: False
+    epsilon: 1
+    clip_grad: False
+    clip_value: 1
+    clip_norm: 1
+    # Data format
+    send_gradient: True
+    # Data loader
+    train_batch_size: 64
+    val_batch_size: 64
+    train_data_shuffle: True
+    val_data_shuffle: False
+
+  model_configs:
+    model_path: "./resources/model/cnn.py"
+    model_name: "CNN"
+    model_kwargs:
+      num_channel: 1
+      num_classes: 10
+      num_pixel: 28
+
+  # comm_configs will be copied to server_configs as well
+  comm_configs:
+    compressor_configs:
+      enable_compression: False
+    # S3 bucket for the Globus Compute clients' model transfer (enabled here).
+    s3_configs:
+      enable_s3: True
+      s3_bucket: "appflx-models"
+
+server_configs:
+  num_clients: 4
+  # FedCompass is asynchronous, which suits a hybrid federation where different
+  # transports (Globus Compute vs TES) may return client updates at different rates.
+  scheduler: "CompassScheduler"
+  scheduler_kwargs:
+    same_init_model: True
+    max_local_steps: 100
+    min_local_steps: 20
+    speed_momentum: 0.9
+    latest_time_factor: 1.5
+  aggregator: "FedCompassAggregator"
+  aggregator_kwargs:
+    weights: "equal"
+    staleness_fn: "polynomial"
+    staleness_fn_kwargs:
+      a: 0.5
+    alpha: 0.9
+  device: "cpu"
+  num_global_epochs: 20
+  logging_output_dirname: "./output"
+  logging_output_filename: "hybrid_result"
+  # File storage for the TES clients (server side).
+  comm_configs:
+    tes_configs:
+      file_storage: "s3"
+      file_storage_kwargs:
+        s3_bucket: "appfl-test-demo-bkt"
+        s3_region: "us-east-1"
+        presigned_url_expiry: 3600
