+++ src/appfl/comm/base/__init__.py
@@ -1,3 +1,11 @@
+from .model_transfer_helper import ModelTransferHelper
 from .base_server_communicator import BaseServerCommunicator
+from .server_comm_backend import ServerCommBackend
+from .server_driven_communicator import ServerDrivenCommunicator
 
-__all__ = ["BaseServerCommunicator"]
+__all__ = [
+    "ModelTransferHelper",
+    "BaseServerCommunicator",
+    "ServerCommBackend",
+    "ServerDrivenCommunicator",
+]
