+++ src/appfl/comm/hybrid/__init__.py
@@ -1,0 +1,3 @@
+from .hybrid_server_communicator import HybridServerCommunicator
+
+__all__ = ["HybridServerCommunicator"]
